#  improve statusLine and StatusMessage by using slice instead of map