We use slice instead of map to store all status messages and status lines. But for the status code bigger than `statusMessageMax(511)`, it still should store them in the dynamic map(`invalidStatusLines`).

@erikdubbelboer PTAL

benchmarks result:
OLD:
```
BenchmarkStatusLine99
BenchmarkStatusLine99-8         278959663                4.51 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine99-8         262258332                4.83 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine99-8         257144448                4.73 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine99-8         251641419                4.80 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200
BenchmarkStatusLine200-8        257207306                4.64 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200-8        257618150                4.71 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200-8        251996420                4.75 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200-8        252930247                4.79 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512
BenchmarkStatusLine512-8        222729294                5.49 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512-8        218682082                5.43 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512-8        218821902                5.48 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512-8        219236438                5.62 ns/op            0 B/op          0 allocs/op
```
NEW:
```
BenchmarkStatusLine99
BenchmarkStatusLine99-8         545864635                2.22 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine99-8         517671476                2.33 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine99-8         505288328                2.31 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine99-8         505923718                2.36 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200
BenchmarkStatusLine200-8        534919548                2.20 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200-8        538629565                2.25 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200-8        523182661                2.30 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine200-8        515232004                2.27 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512
BenchmarkStatusLine512-8        223080207                5.42 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512-8        222268503                5.39 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512-8        216794157                5.47 ns/op            0 B/op          0 allocs/op
BenchmarkStatusLine512-8        216436990                5.57 ns/op            0 B/op          0 allocs/op
```


Co-authored-by: Fenny <fenny@gofiber.io>
Co-authored-by: ReneWerner87 <rene@gofiber.io>