@erikdubbelboer 2020-08-02
Thanks for the PR. I'm looking at this code now and to be honest I think we shouldn't keep invalid status codes in memory. Those shouldn't really happen often enough for this to be worth it (the memory and the code complexity). Can you make `invalidStatusLine()` just return the formatted line and remove `invalidStatusLines`? Thanks.